requirements.txt
mashumaro==3.15
PyYAML==6.0.2
typing_extensions==4.13.1
